feat(cli): optionally remove lock file and also upgrade peerDependencies
Also fixes installation script in case command already taken. release-npm

+++ cli/refresh.js
@@ -3,7 +3,12 @@ import { join } from 'path'
 import rimraf from 'rimraf'
 import { execSync } from 'child_process'
 
+const options = process.argv.splice(2)
+
 rimraf.sync(join(process.cwd(), 'node_modules'))
-rimraf.sync(join(process.cwd(), 'package-lock.json'))
+
+if (options.includes('--lock')) {
+  rimraf.sync(join(process.cwd(), 'package-lock.json'))
+}
 
 execSync('npm install --legacy-peer-deps', { stdio: 'inherit' })
